Move hub date/time adapters to an adapters package. Use 'em in our system tests so ChannelConfig serializes and deserialized correctly via Retrofit2
