+++ src/main/java/com/flightstats/hub/model/ChannelConfig.java
@@ -5,6 +5,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.flightstats.hub.config.properties.ContentProperties;
 import com.flightstats.hub.config.properties.PropertiesLoader;
 import com.flightstats.hub.exception.InvalidRequestException;
+import com.flightstats.hub.model.adapters.HubDateTimeTypeAdapter;
+import com.flightstats.hub.model.adapters.HubDateTypeAdapter;
 import com.flightstats.hub.util.TimeUtil;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
