+++ src/main/java/com/flightstats/hub/model/adapters/HubDateTimeTypeAdapter.java
@@ -1,4 +1,4 @@
-package com.flightstats.hub.model;
+package com.flightstats.hub.model.adapters;
 
 import com.google.gson.JsonSyntaxException;
 import com.google.gson.TypeAdapter;
@@ -12,11 +12,11 @@ import org.joda.time.format.ISODateTimeFormat;
 
 import java.io.IOException;
 
-final class HubDateTimeTypeAdapter extends TypeAdapter<DateTime> {
+public final class HubDateTimeTypeAdapter extends TypeAdapter<DateTime> {
 
     private static DateTimeFormatter dateTimeFormatter = ISODateTimeFormat.dateOptionalTimeParser().withZoneUTC();
 
-    static DateTime deserialize(String json) {
+    public static DateTime deserialize(String json) {
         if (StringUtils.isEmpty(json)) return null;
         DateTime parsed;
         try {
