+++ src/systemTest/java/com/flightstats/hub/system/config/GuiceModule.java
@@ -4,20 +4,26 @@ import com.amazonaws.services.s3.AmazonS3;
 import com.flightstats.hub.clients.callback.CallbackClientFactory;
 import com.flightstats.hub.clients.hub.HubClientFactory;
 import com.flightstats.hub.clients.s3.S3ClientFactory;
+import com.flightstats.hub.model.adapters.HubDateTimeTypeAdapter;
+import com.flightstats.hub.model.adapters.HubDateTypeAdapter;
 import com.flightstats.hub.system.service.CallbackService;
 import com.flightstats.hub.system.service.ChannelService;
 import com.flightstats.hub.system.service.WebhookService;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.name.Names;
 import lombok.extern.slf4j.Slf4j;
 import okhttp3.OkHttpClient;
+import org.joda.time.DateTime;
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
 import retrofit2.converter.scalars.ScalarsConverterFactory;
 
 import javax.inject.Named;
 import javax.inject.Singleton;
+import java.util.Date;
 import java.util.Properties;
 
 @Slf4j
@@ -44,10 +50,14 @@ public class GuiceModule extends AbstractModule {
     @Named("hub")
     @Provides
     public Retrofit retrofitHub(ServiceProperties serviceProperties) {
+        Gson gson = new GsonBuilder()
+                .registerTypeAdapter(Date.class, new HubDateTypeAdapter())
+                .registerTypeAdapter(DateTime.class, new HubDateTimeTypeAdapter())
+                .create();
         return new Retrofit.Builder()
                 .baseUrl(serviceProperties.getHubUrl())
                 .addConverterFactory(ScalarsConverterFactory.create())
-                .addConverterFactory(GsonConverterFactory.create())
+                .addConverterFactory(GsonConverterFactory.create(gson))
                 .client(new OkHttpClient.Builder().build())
                 .build();
     }
